Extract type-related functionality to separate module
With ES6 typeless indexes we want make index types optional. All
type-related functionality in included to index only when it defines
at least 1 type.
