+++ lib/chewy/index.rb
@@ -2,6 +2,7 @@ require 'chewy/search'
 require 'chewy/index/actions'
 require 'chewy/index/aliases'
 require 'chewy/index/settings'
+require 'chewy/index/types'
 require 'chewy/index/specification'
 
 module Chewy
@@ -9,11 +10,9 @@ module Chewy
     include Search
     include Actions
     include Aliases
+    include Types
 
     singleton_class.delegate :client, to: 'Chewy'
-
-    class_attribute :type_hash
-    self.type_hash = {}
 
     class_attribute :_settings
     self._settings = Chewy::Index::Settings.new
@@ -111,80 +110,6 @@ module Chewy
         Chewy.configuration[:prefix]
       end
 
-      # Defines type for the index. Arguments depends on adapter used. For
-      # ActiveRecord you can pass model or scope and options
-      #
-      #   class CarsIndex < Chewy::Index
-      #     define_type Car do
-      #       ...
-      #     end # defines VehiclesIndex::Car type
-      #   end
-      #
-      # Type name might be passed in complicated cases:
-      #
-      #   class VehiclesIndex < Chewy::Index
-      #     define_type Vehicle.cars.includes(:manufacturer), name: 'cars' do
-      #        ...
-      #     end # defines VehiclesIndex::Cars type
-      #
-      #     define_type Vehicle.motocycles.includes(:manufacturer), name: 'motocycles' do
-      #        ...
-      #     end # defines VehiclesIndex::Motocycles type
-      #   end
-      #
-      # For plain objects:
-      #
-      #   class PlanesIndex < Chewy::Index
-      #     define_type :plane do
-      #       ...
-      #     end # defines PlanesIndex::Plane type
-      #   end
-      #
-      # The main difference between using plain objects or ActiveRecord models for indexing
-      # is import. If you will call `CarsIndex::Car.import` - it will import all the cars
-      # automatically, while `PlanesIndex::Plane.import(my_planes)` requires import data to be
-      # passed.
-      #
-      def define_type(target, options = {}, &block)
-        type_class = Chewy.create_type(self, target, options, &block)
-        self.type_hash = type_hash.merge(type_class.type_name => type_class)
-      end
-
-      # Types method has double usage.
-      # If no arguments are passed - it returns array of defined types:
-      #
-      #   UsersIndex.types # => [UsersIndex::Admin, UsersIndex::Manager, UsersIndex::User]
-      #
-      # If arguments are passed it treats like a part of chainable query DSL and
-      # adds types array for index to select.
-      #
-      #   UsersIndex.filters { name =~ 'ro' }.types(:admin, :manager)
-      #   UsersIndex.types(:admin, :manager).filters { name =~ 'ro' } # the same as the first example
-      #
-      def types(*args)
-        if args.present?
-          all.types(*args)
-        else
-          type_hash.values
-        end
-      end
-
-      # Returns defined types names:
-      #
-      #   UsersIndex.type_names # => ['admin', 'manager', 'user']
-      #
-      def type_names
-        type_hash.keys
-      end
-
-      # Returns named type:
-      #
-      #    UserIndex.type('admin') # => UsersIndex::Admin
-      #
-      def type(type_name)
-        type_hash.fetch(type_name) { raise UndefinedType, "Unknown type in #{name}: #{type_name}" }
-      end
-
       # Used as a part of index definition DSL. Defines settings:
       #
       #   class UsersIndex < Chewy::Index
@@ -211,7 +136,7 @@ module Chewy
       # Returns list of public class methods defined in current index
       #
       def scopes
-        public_methods - Chewy::Index.public_methods
+        public_methods - Chewy::Index.public_methods - Types::ADDITIONAL_METHODS
       end
 
       def settings_hash
@@ -219,6 +144,7 @@ module Chewy
       end
 
       def mappings_hash
+        return {} unless has_types?
         mappings = types.map(&:mappings_hash).inject(:merge)
         mappings.present? ? {mappings: mappings} : {}
       end
