+++ lib/chewy/index/actions.rb
@@ -136,6 +136,7 @@ module Chewy
         %i[import import!].each do |method|
           class_eval <<-METHOD, __FILE__, __LINE__ + 1
             def #{method}(*args)
+              return unless has_types?
               options = args.extract_options!
               if args.one? && type_names.one?
                 objects = {type_names.first.to_sym => args.first}
